Update: Moved validation back to form for testing & model creation
