+++ job_tracker/app/models/job_descriptions/job_description_form.rb
@@ -3,6 +3,15 @@ class JobDescriptions::JobDescriptionForm
 
   attr_accessor :job_description, :description, :requirements, 
     :job_classification, :min_salary, :max_salary, :job_application_id
+
+    validates   :description, :requirements, :job_classification,
+    :min_salary, :max_salary, presence: true
+
+  validates :job_classification, 
+    format: { with: /\A[a-zA-Z\s]+\z/ },  unless: -> { job_classification.blank? }
+
+  validates :min_salary, :max_salary, numericality: { greater_than: 0 }
+  validate :min_salary_less_than_max_salary
     
 
   def initialize(job_description:)
@@ -25,11 +34,28 @@ class JobDescriptions::JobDescriptionForm
   end
 
   def update!(job_description_params)
-    job_description.attributes = job_description_params #TODO: figure out how to set attributes
-    if job_description.valid?
+    @description = job_description_params[:description]
+    @requirements = job_description_params[:requirements]
+    @job_classification = job_description_params[:job_classification]
+    @min_salary = job_description_params[:min_salary]
+    @max_salary = job_description_params[:max_salary]
+  
+    if valid?
       job_description.update!(job_description_params)
     else
+      job_description.errors.copy!(self.errors)
       false
     end
   end
+
+  private
+  
+  def min_salary_less_than_max_salary
+    return unless min_salary.present? && max_salary.present?
+
+    if min_salary > max_salary
+      errors.add(:min_salary, 'must be less than max salary')
+    end
+  end
+
 end
