+++ job_tracker/spec/support/factory_bot.rb
@@ -1,3 +1,5 @@
+require 'factory_bot'
+
 RSpec.configure do |config|
     config.include FactoryBot::Syntax::Methods
   end
